continue refactoring contract logic
