+++ src/clients/contracts/contractHelpers.ts
@@ -1,24 +1,24 @@
 import Web3 from 'web3';
 
 import { getWeb3NoAccount } from 'clients/web3';
-import bep20Abi from 'constants/contractAbis/bep20.json';
-import comptrollerAbi from 'constants/contractAbis/comptroller.json';
-import interestModelAbi from 'constants/contractAbis/interestModel.json';
-import oracleAbi from 'constants/contractAbis/oracle.json';
-import vaiTokenAbi from 'constants/contractAbis/vaiToken.json';
-import vaiUnitrollerAbi from 'constants/contractAbis/vaiUnitroller.json';
-import vaiVaultAbi from 'constants/contractAbis/vaiVault.json';
-import xvsVaultStoreAbi from 'constants/contractAbis/xvsVaultStore.json';
-import xvsVaultAbi from 'constants/contractAbis/xvsVault.json';
-import vbepAbi from 'constants/contractAbis/vbep.json';
-import vbnbAbi from 'constants/contractAbis/vbnb.json';
-import xvsAbi from 'constants/contractAbis/xvs.json';
-import venusLensAbi from 'constants/contractAbis/venusLens.json';
-import governorBravoAbi from 'constants/contractAbis/governorBravoDelegate.json';
-import xvsVestingAbi from 'constants/contractAbis/xvsVesting.json';
-import vrtConverterAbi from 'constants/contractAbis/vrtConverter.json';
-import vrtTokenAbi from 'constants/contractAbis/vrtToken.json';
-import vrtVaultAbi from 'constants/contractAbis/vrtVault.json';
+import bep20Abi from 'constants/contracts/abis/bep20.json';
+import comptrollerAbi from 'constants/contracts/abis/comptroller.json';
+import interestModelAbi from 'constants/contracts/abis/interestModel.json';
+import oracleAbi from 'constants/contracts/abis/oracle.json';
+import vaiTokenAbi from 'constants/contracts/abis/vaiToken.json';
+import vaiUnitrollerAbi from 'constants/contracts/abis/vaiUnitroller.json';
+import vaiVaultAbi from 'constants/contracts/abis/vaiVault.json';
+import xvsVaultStoreAbi from 'constants/contracts/abis/xvsVaultStore.json';
+import xvsVaultAbi from 'constants/contracts/abis/xvsVault.json';
+import vbepAbi from 'constants/contracts/abis/vbep.json';
+import vbnbAbi from 'constants/contracts/abis/vbnb.json';
+import xvsAbi from 'constants/contracts/abis/xvs.json';
+import venusLensAbi from 'constants/contracts/abis/venusLens.json';
+import governorBravoAbi from 'constants/contracts/abis/governorBravoDelegate.json';
+import xvsVestingAbi from 'constants/contracts/abis/xvsVesting.json';
+import vrtConverterAbi from 'constants/contracts/abis/vrtConverter.json';
+import vrtTokenAbi from 'constants/contracts/abis/vrtToken.json';
+import vrtVaultAbi from 'constants/contracts/abis/vrtVault.json';
 import { getContractAddress, getToken, getVBepToken } from 'utilities';
 
 const getContract = (abi: $TSFixMe, address: $TSFixMe, web3Contract: Web3) => {
@@ -27,7 +27,7 @@ const getContract = (abi: $TSFixMe, address: $TSFixMe, web3Contract: Web3) => {
 };
 
 export const getVaiTokenContract = (web3: Web3) =>
-  getContract(vaiTokenAbi, getContractAddress('vaiToken'), web3);
+  getContract(vaiTokenAbi, getContractAddress('vai'), web3);
 
 export const getVaiUnitrollerContract = (web3: Web3) =>
   getContract(vaiUnitrollerAbi, getContractAddress('vaiUnitroller'), web3);
@@ -79,7 +79,7 @@ export const getVrtConverterProxyContract = (web3: Web3) =>
   getContract(vrtConverterAbi, getContractAddress('vrtConverterProxy'), web3);
 
 export const getVrtTokenContract = (web3: Web3) =>
-  getContract(vrtTokenAbi, getContractAddress('vrtToken'), web3);
+  getContract(vrtTokenAbi, getContractAddress('vrt'), web3);
 
 // VRT vault
 export const getVrtVaultProxyContract = (web3: Web3) =>
