+++ src/constants/contracts/addresses/main.json
@@ -1,12 +1,4 @@
 {
-  "vrtToken": {
-    "56": "0x5f84ce30dc3cf7909101c69086c50de191895883",
-    "97": "0x331F639B4F3CF6E391CD244e0b5027C5968Ec325"
-  },
-  "vaiToken": {
-    "56": "0x4BD17003473389A42DAF6a0a729f6Fdb328BbBd7",
-    "97": "0x5fFbE5302BadED40941A403228E6AD03f93752d9"
-  },
   "vaiVault": {
     "56": "0x0667Eed0a0aAb930af74a3dfeDD263A73994f216",
     "97": "0x7Db4f5cC3bBA3e12FF1F528D2e3417afb0a57118"
