+++ src/constants/contracts/addresses/tokens.json
@@ -1,0 +1,10 @@
+{
+  "vrt": {
+    "56": "0x5f84ce30dc3cf7909101c69086c50de191895883",
+    "97": "0x331F639B4F3CF6E391CD244e0b5027C5968Ec325"
+  },
+  "vai": {
+    "56": "0x4BD17003473389A42DAF6a0a729f6Fdb328BbBd7",
+    "97": "0x5fFbE5302BadED40941A403228E6AD03f93752d9"
+  }
+}
