+++ src/utilities/common.ts
@@ -2,7 +2,7 @@ import BigNumber from 'bignumber.js';
 import { ethers } from 'ethers';
 import commaNumber from 'comma-number';
 
-import { TOKENS } from 'constants/tokenContracts';
+import { TOKENS } from 'constants/tokens';
 import { getToken, getVBepToken } from 'utilities';
 import { TokenSymbol } from 'types';
 import getContractAddress from './getContractAddress';
@@ -49,7 +49,7 @@ export const addToken = async ({
   let tokenDecimals = 18;
   let tokenImage = '';
   if (asset === 'vai') {
-    tokenAddress = getContractAddress('vaiToken');
+    tokenAddress = getContractAddress('vai');
     tokenSymbol = 'VAI';
     tokenDecimals = 18;
     tokenImage = `${window.location.origin}/coins/vai.svg`;
